fixed minor bugs and issues #43 #42 #23
