+++ ACDD_conf.yaml
@@ -48,6 +48,8 @@ cmor:
     variable_to_process: prw 
     var_subset: !!bool False
     var_subset_list: ''
+    # if subhr data is included specify actual frequency as ##min
+    subhr: 10min
     # Mappings, vocab and tables settings
     # default=data/dreq/cmvme_all_piControl_3_3.csv
     # Leave as set unless publishing for CMIP6
@@ -78,6 +80,8 @@ cmor:
     # queue and memory (GB) per CPU (depends on queue) 
     queue: hugemem
     mem_per_cpu: 30 
+    # walltime in "hh:mm:ss"
+    walltime: '8:00:00'
     mode: custom
 #
 # Global attributes: these will be added to each files comment unwanted ones
